pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-applier"
version = "2.0.0"
description = "AI-powered automated job application tool with multi-platform support"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "patchright>=1.50",
    "playwright>=1.40",
    "playwright-stealth>=1.0.6",
    "pdfplumber>=0.10",
    "python-docx>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pyinstaller>=6.0",
]

[project.scripts]
auto-applier = "auto_applier.main:cli"

[project.gui-scripts]
auto-applier-gui = "auto_applier.gui.wizard:launch_wizard"

[tool.setuptools.packages.find]
include = ["auto_applier*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
